Split local git operations from github operations
Previously the functions which identify the changes to make to the target
repository and which apply them to get a patch were intermingled with the
functions interacting with github.  This meant that testing the file and
patching operations required github credentials.

This patch separates out the file operations into their own command,
`patch-target.sh`, which is called with the relevant parameters by
`entrypoint.sh`.

It also makes some additions to the error handling in the
`update-from-template` command to make debugging a little easier.

Fixes #48.